Added some C header files to have the definition of C functions Removed some unused local variables
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@9867 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/parsing/eiffel/byte.c
@@ -159,7 +159,7 @@ long ca_suint32(void)
 
 void ca_store(char *area, long int siz, FILE *fil)
 {
-	if (fwrite(area, sizeof(char), siz, fil) != siz)
+	if (fwrite(area, sizeof(char), siz, fil) != (size_t) siz)
 		xraise(EN_IO);
 }
